refactor(scaffold): consolidate Preset/SectionTab/SceneTab onto TapButton (#156)
Lift the verbatim-duplicate machinery (sphere mesh + ray-hit + press-flash
+ haptic + yaw-billboard label + emissive priority + dispose) shared by
Preset (#46), SectionTab (#57), and SceneTab (#150) into a new
`scaffold/ui/TapButton` base. Each primitive becomes a thin subclass that
calls `super()` with its visual identity (radius, base/hover/press/active
emissive, label font/offset/anchor, group-name prefix); Preset keeps its
`values` / `linearValues` extra fields.

`isActive` is gated on `activeEmissive !== undefined` so one-shot
affordances (Preset, which omits `activeEmissive`) can never report a
sticky-active state — if the button can't show active, it can't be active.

Pure refactor — no user-visible change. Net: -197 lines.
